move test to allinone
